Improve db sync dependencies
This commit adds a couple of dependencies related to db sync.

Previously, it was possible that the nova-scheduler could be started
before the db was synced. This would result in the service failing.
This could only happen if none of the nova::manage::* defines were used.

This commit adds a few extra dependencies to ensure that the initial
db sync occurs before any services are started and also that the nova-db-sync
call is only made after the nova-command package is installed and nova
is configured to connect to the db.